Struct phf::ordered_map::OrderedMap

pub struct OrderedMap<K: 'static, V: 'static> {
    // some fields omitted
}

An order-preserving immutable map constructed at compile time.

Unlike a Map, iteration order is guaranteed to match the definition order.

Note

The fields of this struct are public so that they may be initialized by the phf_ordered_map! macro and code generation. They are subject to change at any time and should never be accessed directly.

Methods

impl<K, V> OrderedMap<K, V>

fn len(&self) -> usize

Returns the number of entries in the Map.

fn is_empty(&self) -> bool

Returns true if the Map is empty.

fn get<T: ?Sized>(&self, key: &T) -> Option<&V> where T: Eq + PhfHash, K: Borrow<T>

Returns a reference to the value that key maps to.

fn get_key<T: ?Sized>(&self, key: &T) -> Option<&K> where T: Eq + PhfHash, K: Borrow<T>

Returns a reference to the map's internal static instance of the given key.

This can be useful for interning schemes.

fn contains_key<T: ?Sized>(&self, key: &T) -> bool where T: Eq + PhfHash, K: Borrow<T>

Determines if key is in the Map.

fn get_index<T: ?Sized>(&self, key: &T) -> Option<usize> where T: Eq + PhfHash, K: Borrow<T>

Returns the index of the key within the list used to initialize the ordered map.

fn index(&self, index: usize) -> Option<(&K, &V)>

Returns references to both the key and values at an index within the list used to initialize the ordered map. See .get_index(key).

fn get_entry<T: ?Sized>(&self, key: &T) -> Option<(&K, &V)> where T: Eq + PhfHash, K: Borrow<T>

Like get, but returns both the key and the value.

fn entries<'a>(&'a self) -> Entries<'a, K, V>

Returns an iterator over the key/value pairs in the map.

Entries are returned in the same order in which they were defined.

fn keys<'a>(&'a self) -> Keys<'a, K, V>

Returns an iterator over the keys in the map.

Keys are returned in the same order in which they were defined.

fn values<'a>(&'a self) -> Values<'a, K, V>

Returns an iterator over the values in the map.

Values are returned in the same order in which they were defined.