Struct phf::Map

pub struct Map<K: 'static, V: 'static> { /* fields omitted */ }

An immutable map constructed at compile time.

Note

The fields of this struct are public so that they may be initialized by the phf_map! macro and code generation. They are subject to change at any time and should never be accessed directly.

Implementations

impl<K, V> Map<K, V>

pub fn is_empty(&self) -> bool

Returns true if the Map is empty.

pub fn len(&self) -> usize

Returns the number of entries in the Map.

pub fn contains_key<T: ?Sized>(&self, key: &T) -> bool where
T: Eq + PhfHash,
K: Borrow<T>,

Determines if key is in the Map.

pub fn get<T: ?Sized>(&self, key: &T) -> Option<&V> where
T: Eq + PhfHash,
K: Borrow<T>,

Returns a reference to the value that key maps to.

pub fn get_key<T: ?Sized>(&self, key: &T) -> Option<&K> where
T: Eq + PhfHash,
K: Borrow<T>,

Returns a reference to the map's internal static instance of the given key.

This can be useful for interning schemes.

pub fn get_entry<T: ?Sized>(&self, key: &T) -> Option<(&K, &V)> where
T: Eq + PhfHash,
K: Borrow<T>,

Like get, but returns both the key and the value.

pub fn entries<'a>(&'a self) -> Entries<'a, K, V>

Returns an iterator over the key/value pairs in the map.

Entries are returned in an arbitrary but fixed order.

pub fn keys<'a>(&'a self) -> Keys<'a, K, V>

Returns an iterator over the keys in the map.

Keys are returned in an arbitrary but fixed order.

pub fn values<'a>(&'a self) -> Values<'a, K, V>

Returns an iterator over the values in the map.

Values are returned in an arbitrary but fixed order.