Struct phf::OrderedSet

pub struct OrderedSet<T: 'static> { /* private fields */ }

An order-preserving immutable set constructed at compile time.

Unlike a Set, iteration order is guaranteed to match the definition order.

Note

The fields of this struct are public so that they may be initialized by the phf_ordered_set! macro and code generation. They are subject to change at any time and should never be accessed directly.

Implementations

impl<T> OrderedSet<T>

pub const fn len(&self) -> usize

Returns the number of elements in the OrderedSet.

pub const fn is_empty(&self) -> bool

Returns true if the OrderedSet contains no elements.

pub fn get_key<U: ?Sized>(&self, key: &U) -> Option<&T> where
U: Eq + PhfHash,
T: PhfBorrow<U>,

Returns a reference to the set’s internal static instance of the given key.

This can be useful for interning schemes.

pub fn get_index<U: ?Sized>(&self, key: &U) -> Option<usize> where
U: Eq + PhfHash,
T: PhfBorrow<U>,

Returns the index of the key within the list used to initialize the ordered set.

pub fn index(&self, index: usize) -> Option<&T>

Returns a reference to the key at an index within the list used to initialize the ordered set. See .get_index(key).

pub fn contains<U: ?Sized>(&self, value: &U) -> bool where
U: Eq + PhfHash,
T: PhfBorrow<U>,

Returns true if value is in the OrderedSet.

pub fn iter(&self) -> Iter<'_, T>

Returns an iterator over the values in the set.

Values are returned in the same order in which they were defined.

impl<T> OrderedSet<T> where
T: Eq + PhfHash + PhfBorrow<T>,

pub fn is_disjoint(&self, other: &OrderedSet<T>) -> bool

Returns true if other shares no elements with self.

pub fn is_subset(&self, other: &OrderedSet<T>) -> bool

Returns true if other contains all values in self.

pub fn is_superset(&self, other: &OrderedSet<T>) -> bool

Returns true if self contains all values in other.