phf

Trait PhfHash

pub trait PhfHash {
    // Required method
    fn phf_hash<H>(&self, state: &mut H)
       where H: Hasher;

    // Provided method
    fn phf_hash_slice<H>(data: &[Self], state: &mut H)
       where H: Hasher,
             Self: Sized { ... }
}

A trait implemented by types which can be used in PHF data structures.

This differs from the standard library’s Hash trait in that PhfHash’s results must be architecture independent so that hashes will be consistent between the host and target when cross compiling.

Required Methods

fn phf_hash<H>(&self, state: &mut H)
where H: Hasher,

Feeds the value into the state given, updating the hasher as necessary.

Provided Methods

fn phf_hash_slice<H>(data: &[Self], state: &mut H)
where H: Hasher, Self: Sized,

Feeds a slice of this type into the state provided.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.