Crate phf

Compile time optimized maps and sets.

PHF data structures can be generated via the syntax extensions in the phf_macros crate or via code generation in the phf_codegen crate. See the documentation of those crates for more details.

Modules

map
An immutable map constructed at compile time.

ordered_map
An order-preserving immutable map constructed at compile time.

ordered_set
An order-preserving immutable set constructed at compile time.

set
An immutable set constructed at compile time.

Structs

Map
An immutable map constructed at compile time.

OrderedMap
An order-preserving immutable map constructed at compile time.

OrderedSet
An order-preserving immutable set constructed at compile time.

Set
An immutable set constructed at compile time.

Traits

PhfHash
A trait implemented by types which can be used in PHF data structures.