Crate phf

Compile time optimized maps and sets.

Keys can be string literals, byte string literals, byte literals, char literals, or any of the fixed-size isizeegral types.

Examples

#![feature(plugin)]
#![plugin(phf_macros)]

extern crate phf;

#[derive(Clone)]
pub enum Keyword {
    Loop,
    Continue,
    Break,
    Fn,
    Extern,
}

static KEYWORDS: phf::Map<&'static str, Keyword> = phf_map! {
    "loop" => Keyword::Loop,
    "continue" => Keyword::Continue,
    "break" => Keyword::Break,
    "fn" => Keyword::Fn,
    "extern" => Keyword::Extern,
};

pub fn parse_keyword(keyword: &str) -> Option<Keyword> {
    KEYWORDS.get(keyword).cloned()
}

Modules

map
An immutable map constructed at compile time.

ordered_map
An order-preserving immutable map constructed at compile time.

ordered_set
An order-preserving immutable set constructed at compile time.

set
An immutable set constructed at compile time.

Structs

Map
An immutable map constructed at compile time.

OrderedMap
An order-preserving immutable map constructed at compile time.

OrderedSet
An order-preserving immutable set constructed at compile time.

Set
An immutable set constructed at compile time.

Traits

PhfHash
A trait implemented by types which can be used in PHF data structures