Struct phf::Set

pub struct Set<T: 'static> { /* fields omitted */ }

An immutable set constructed at compile time.

Note

The fields of this struct are public so that they may be initialized by the phf_set! macro and code generation. They are subject to change at any time and should never be accessed directly.

Methods

impl<T> Set<T>

fn len(&self) -> usize

Returns the number of elements in the Set.

fn is_empty(&self) -> bool

Returns true if the Set contains no elements.

fn get_key<U: ?Sized>(&self, key: &U) -> Option<&T> where U: Eq + PhfHash, T: Borrow<U>

Returns a reference to the set's internal static instance of the given key.

This can be useful for interning schemes.

fn contains<U: ?Sized>(&self, value: &U) -> bool where U: Eq + PhfHash, T: Borrow<U>

Returns true if value is in the Set.

fn iter<'a>(&'a self) -> Iter<'a, T>

Returns an iterator over the values in the set.

Values are returned in an arbitrary but fixed order.

impl<T> Set<T> where T: Eq + PhfHash

fn is_disjoint(&self, other: &Set<T>) -> bool

Returns true if other shares no elements with self.

fn is_subset(&self, other: &Set<T>) -> bool

Returns true if other contains all values in self.

fn is_superset(&self, other: &Set<T>) -> bool

Returns true if self contains all values in other.